Fix to TinyMCE relative path handling
The TinyMCE library converts iframe links that go to the same service from
absolute to relative. Currently, it only checks the host and protocol,
meaning that if two services are running on the same host, but on different
ports, it believes links between them to be to the same service and
so converts them to relative. This causes an issue when an external
tool running on the same host as the Canvas installation (common in dev/test
setups) creates an iframe through a TinyMCE editor window. This fix changes
the absolute-to-relative conversion code to check for port number.